fix: make options optional and deep merge options
To allow passing partial configs, make all options properties
optional.
